Tidying, preventing potential null ref exceptions

+++ ICD.Connect.Lighting.CrestronPro/Shades/AbstractShadeBaseAdapter.cs
@@ -130,7 +130,7 @@ namespace ICD.Connect.Lighting.CrestronPro.Shades
 
 		public float GetPosition()
 		{
-			return MathUtils.MapRange(0, (float)ushort.MaxValue, 0, 1, (float)Shade.PositionFeedback.UShortValue);
+			return MathUtils.MapRange(0, ushort.MaxValue, 0, 1, (float)Shade.PositionFeedback.UShortValue);
 		}
 
 		public void SetPosition(float position)
